--- monew/src/main/java/com/codeit/team2/monew/module/domain/user/dto/UserUpdateRequest.java
@@ -1,8 +1,0 @@
-package com.codeit.team2.monew.module.domain.user.dto;
-
-public record UserUpdateRequest(
-
-    String nickname
-) {
-
-}
