+++ monew/src/main/java/com/codeit/team2/monew/module/domain/user/dto/request/UserRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.codeit.team2.monew.module.domain.user.dto;
+package com.codeit.team2.monew.module.domain.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
