+++ monew/src/main/java/com/codeit/team2/monew/module/domain/user/dto/request/UserUpdateRequest.java
@@ -1,0 +1,8 @@
+package com.codeit.team2.monew.module.domain.user.dto.request;
+
+public record UserUpdateRequest(
+
+    String nickname
+) {
+
+}
