+++ monew/src/main/java/com/codeit/team2/monew/module/domain/user/service/UserService.java
@@ -1,8 +1,8 @@
 package com.codeit.team2.monew.module.domain.user.service;
 
-import com.codeit.team2.monew.module.domain.user.dto.UserDto;
-import com.codeit.team2.monew.module.domain.user.dto.UserRegisterRequest;
-import com.codeit.team2.monew.module.domain.user.dto.UserUpdateRequest;
+import com.codeit.team2.monew.module.domain.user.dto.request.UserRegisterRequest;
+import com.codeit.team2.monew.module.domain.user.dto.request.UserUpdateRequest;
+import com.codeit.team2.monew.module.domain.user.dto.response.UserDto;
 import java.util.UUID;
 
 public interface UserService {
